[Xamarin.Android.Build.Tasks] remove usage of $(TFV) in _ResolveSdks
As we migrate to .NET 5, we need a way for short-form MSBuild projects
to run a different set of MSBuild tasks/targets within the codebase.

At the top level `Xamarin.Android.Sdk.targets` will be imported
instead of `Xamarin.Android.CSharp.targets`. We can define
`$(UsingAndroidNETSdk)` as `true`, which is similar to
`$(UsingMicrosoftNETSdk)` already present for short-form MSBuild
projects.

Based on `$(UsingAndroidNETSdk)`:

* Existing long-form projects can import
  `Xamarin.Android.Legacy.targets`
* New short-form project can import other new files as needed.

The first two MSBuild tasks to change are:

* `<ValidateJavaVersion/>`
* `<ResolveAndroidTooling/>`

Both of these tasks rely on `$(TargetFrameworkVersion)`; however, we
can make versions of these tasks that don't need to use
`$(TargetFrameworkVersion)` at all.

`<ValidateJavaVersion/>` can just require Java 1.8 going forward.

`<ResolveAndroidTooling/>` will have any logic around
`$(TargetFrameworkVersion)` or `$(AndroidUseLatestPlatformSdk)` and
just targets the latest supported Android API level.

To make the two versions of these tasks:

* The core code of these tasks will remain.
* A new `Tasks\Legacy` folder and `Xamarin.Android.Tasks.Legacy`
  namespace will be created.
* Any "legacy" code can be moved into these classes that will inherit
  from the core tasks.
* Any MSBuild targets using these tasks will be included in
  `Xamarin.Android.Legacy.targets` via `$(UsingAndroidNETSdk)` as well
  as any `<UsingTask/>` declarations.

~~ Binding Projects ~~

Since Binding projects have their own calls to
`<ValidateJavaVersion/>` and `<ResolveAndroidTooling/>`, now is the
time to consolidate this. We had some duplicated code here.

I moved the common targets into `Xamarin.Android.Tooling.targets`, all
projects will now use the `_ResolveSdks` target from this file.

I had to introduce a private property:

    <_AndroidIsBindingProject>True</_AndroidIsBindingProject>

We can use this to detect if the current project is a binding project
or not. This may evolve as we fully understand what needs to happen
for .NET 5.
