+++ src/Xamarin.Android.Build.Tasks/MSBuild/Xamarin/Android/Xamarin.Android.Aapt.targets
@@ -16,7 +16,7 @@ Copyright (C) 2019  Microsoft Corporation. All rights reserved.
 
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
-<Target Name="_InjectAaptDependencies" AfterTargets="_ResolveSdks" Condition=" '$(_AndroidUseAapt2)' != 'True' ">
+<Target Name="_InjectAaptDependencies" Condition=" '$(_AndroidUseAapt2)' != 'True' ">
   <PropertyGroup>
     <_UpdateAndroidResgenInputs>
       $(_UpdateAndroidResgenInputs);
